+++ examples/custom-node/src/primitives/mod.rs
@@ -9,8 +9,8 @@ pub use tx::*;
 
 pub mod tx_type;
 pub use tx_type::*;
-pub mod tx_custom;
-pub use tx_custom::*;
+pub mod tx_payment;
+pub use tx_payment::*;
 
 use reth_ethereum::primitives::NodePrimitives;
 use reth_op::OpReceipt;
